+++ init.ps1
@@ -16,7 +16,10 @@ If BuildMaster is already installed, nothing happens.
 #>
 [CmdletBinding()]
 param(
-    [String] $SqlServerName
+    [String] $SqlServerName,
+
+    [Parameter(Mandatory)]
+    [String] $Version
 )
 
 #Requires -RunAsAdministrator
@@ -25,14 +28,14 @@ Set-StrictMode -Version 'Latest'
 $ErrorActionPreference = 'Stop'
 $InformationPreference = 'Continue'
 
-Import-Module -Name (Join-Path -Path $PSScriptRoot -ChildPath 'PSModules\Carbon.Windows.Installer') -Force -Verbose:$false
 Import-Module -Name 'Microsoft.PowerShell.Archive' -Verbose:$false
 
-# When updating the version, it's a good time to check if bugs in the API have been fixed. Search all the tests for
-# "-Skip", remove the "-Skip" flag and run tests.
-$version = '24.0.10'
+$psModulesDirPath = Join-Path -Path $PSScriptRoot -ChildPath 'PSModules' -Resolve
+Import-Module -Name (Join-Path -Path $psModulesDirPath -ChildPath 'Yodel' -Resolve) `
+              -Function @('Connect-YDatabase', 'Invoke-YMsSqlCommand') `
+              -Verbose:$false
 
-Write-Information -Message "Testing BuildMaster ${version}."
+Write-Information -Message "Installing BuildMaster ${Version}."
 
 $outputPath = Join-Path -Path $PSScriptRoot -ChildPath '.output'
 if (-not (Test-Path -Path $outputPath))
@@ -64,7 +67,7 @@ if ($runningUnderAppVeyor)
 }
 elseif (-not $SqlServerName)
 {
-    Import-Module -Name (Join-Path -Path $PSScriptRoot -ChildPath 'PSModules\SqlServer') -Force
+    Import-Module -Name (Join-Path -Path $psModulesDirPath -ChildPath 'SqlServer') -Force -Verbose:$false
 
     $machineSqlPath = Join-Path -Path 'SQLSERVER:\SQL' -ChildPath ([Environment]::MachineName)
     $sqlServer =
@@ -93,9 +96,123 @@ elseif (-not $SqlServerName)
 Write-Information "Using SQL Server ${SqlServerName}."
 
 # Free edition license
-& $hubPath 'install' `
-           "BuildMaster:$($version)" `
-           --ConnectionString="Server=${SqlServerName}; ${dbCredentials}" `
-           --LicenseKey=C2G2G010-GC100V-7M8X70-0QC1GFNK-H98U
+$licenseKey = 'C2G2G010-GC100V-7M8X70-0QC1GFNK-H98U'
+$connString = "Server=${SqlServerName}; ${dbCredentials}"
+Write-Verbose "Connection String  ${connString}"
 
-Get-Service -Name 'Inedo*' | Start-Service
+& $hubPath 'install' "BuildMaster:${Version}" --ConnectionString="${connString}" --LicenseKey="${licenseKey}"
+
+$bmSvcs = Get-Service -Name 'INEDOBM*'
+
+$bmSvcs | Start-Service
+
+$bmReconfigured = $false
+
+$conn = Connect-YDatabase -Provider ([Data.SqlClient.SqlClientFactory]::Instance) `
+                          -ConnectionString "${connString}; Database=BuildMaster"
+
+$query = 'select name from sys.procedures where name=''ApiKeys_GetApiKeyByName'''
+while ($true)
+{
+    Write-Information 'Waiting for BuildMaster database.'
+    try
+    {
+        if (Invoke-YMsSqlCommand -Connection $conn -Text $query -AsScalar)
+        {
+            break
+        }
+    }
+    catch
+    {
+        Write-Warning $_
+    }
+    Start-Sleep -Seconds 1
+}
+
+try
+{
+    $apiKey = Get-Content -Path (Join-Path -Path $PSScriptRoot -ChildPath '.apikey') -ReadCount 1
+
+    $keyExists = Invoke-YMsSqlCommand -Connection $conn `
+                                      -Text '[dbo].[ApiKeys_GetApiKeyByName]' `
+                                      -Type StoredProcedure `
+                                      -AsScalar `
+                                      -Parameter @{ '@ApiKey_Text' = $apiKey }
+    if (-not $keyExists)
+    {
+        $apiKeyDescription = 'BuildMasterAutomation API Key'
+        $apiKeyConfig = @'
+<Inedo.BuildMaster.ApiKeys.ApiKey Assembly="BuildMaster">
+    <Properties AllowNativeApi="True" CanViewInfrastructure="True" CanUpdateInfrastructure="True" AllowVariablesManagementApi="True" AllowReleaseAndPackageDeploymentApi="True" />
+</Inedo.BuildMaster.ApiKeys.ApiKey>
+'@
+
+        Write-Information 'Inserting test API key into BuildMaster database.'
+        Invoke-YMsSqlCommand -Connection $conn `
+                             -Text '[dbo].[ApiKeys_CreateOrUpdateApiKey]' `
+                             -Type StoredProcedure `
+                             -Parameter @{
+                                    '@ApiKey_Text' = $apiKey;
+                                    '@ApiKey_Description' = $apiKeyDescription;
+                                    '@ApiKey_Configuration' = $apiKeyConfig
+                                } `
+                             -NonQuery | Out-Null
+
+        $bmReconfigured = $true
+    }
+
+    $currentLicense = Invoke-YMsSqlCommand -Connection $conn `
+                                           -Text '[dbo].[Configuration_GetValue]' `
+                                           -Type StoredProcedure `
+                                           -Parameter @{ '@Key_Name' = 'Licensing.Key' }
+    if (-not $currentLicense)
+    {
+        Write-Information 'Adding license key to BuildMaster database.'
+        Invoke-YMsSqlCommand -Connection $conn `
+                             -Text '[dbo].[Configuration_SetValue]' `
+                             -Type StoredProcedure `
+                             -Parameter @{
+                                    '@Key_Name' = 'Licensing.Key'
+                                    '@Value_Text' = $licenseKey
+                                } `
+                             -NonQuery | Out-Null
+
+        $bmReconfigured = $true
+    }
+}
+finally
+{
+    $conn.Close()
+}
+
+if ($bmReconfigured)
+{
+    Write-Information 'Restarting BuildMaster services.'
+    $bmSvcs | Restart-Service
+}
+
+Import-Module -Name (Join-Path -Path $PSScriptRoot -ChildPath 'BuildMasterAutomation' -Resolve) -Force -Verbose:$false
+
+$session = New-BMSession -Url "http://$([Environment]::MachineName):8622/" -ApiKey $apiKey
+
+# Before BuildMaster is activated, the APIs return the licensing page HTML.
+do
+{
+    Write-Information 'Waiting for BuildMaster activation.'
+    try
+    {
+        $result = Invoke-BMRestMethod -Session $session -Name 'infrastructure/servers/list'
+        if ($result -isnot [String])
+        {
+            break
+        }
+        $result | Out-String | Write-Verbose
+    }
+    catch
+    {
+        Write-Warning $_
+    }
+
+    Start-Sleep -Seconds 1
+}
+while ($true)